% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_latest_version.R
\name{is_latest_version}
\alias{is_latest_version}
\title{Check the Latest Version of a Synapse ID}
\usage{
is_latest_version(syn_id, version = NULL, authToken = NULL, verbose = TRUE)
}
\arguments{
\item{syn_id}{a string specifying the Synapse ID (e.g. "syn123") or the ID
plus the version number (e.g. "syn123.5"). If the version number is given,
it must follow the format '\verb{<id>.<version>}' and the version must be a
number.}

\item{version}{(optional) If no version is specified in \code{syn_id}, this should
be a number or something coercible to a number, specifying the version.
Default: \code{NULL}.}

\item{authToken}{(optional) A string containing the user's Synapse auth
token, which is used as the \code{authToken} argument to \link[synapser:synLogin]{synapser::synLogin}.
If \code{NULL}, \code{synLogin} will attempt to read the user's credentials from
stored configuration. Default: \code{NULL}.}

\item{verbose}{(optional) if \code{TRUE}, print out whether the given version is
the latest version or not, before returning. If \code{FALSE}, nothing is
printed. Default: \code{TRUE}.}
}
\value{
a boolean variable indicating whether the given version is the latest
version on Synapse, or not.
}
\description{
Given a Synapse ID and a file version, this function checks if that version
is the latest version available on Synapse.
}
\examples{
\dontrun{
is_latest_version("syn123", version = 4)
is_latest_version("syn123.4")
}
}
