% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_synStore.R
\name{safe_synStore}
\alias{safe_synStore}
\title{Safe Synapse Store}
\usage{
safe_synStore(
  filename,
  parent_id,
  set_annotations = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{filename}{the path to the file. This can be either a local path (e.g.
"data/file.txt") or an absolute path (e.g. ~/Users/user1/data/file.txt").}

\item{parent_id}{the Synapse ID of the folder on Synapse where this file
should be stored}

\item{set_annotations}{(optional) the \code{set_annotations} argument to
\link{synStore}. Default: \code{FALSE}, which will not delete annotations on the
file.}

\item{verbose}{(optional) If \code{TRUE}, the function will print a message if the
file matches what is on Synapse and isn't uploaded. If \code{FALSE}, no message
is printed. Default: \code{TRUE}.}

\item{...}{Other optional arguments to \link{synStore}, such as \code{used} or
\code{executed}.}
}
\value{
a synapser::File object
}
\description{
This is a wrapper around \link[synapser:synStore]{synapser::synStore} that "safely" uploads a file to
Synapse without erasing annotations or version comments. This function
assumes the user has already logged in with \link{synLogin}.
}
\details{
By default, \code{synStore} deletes any annotations and version comments on the
file even if the file to upload is identical to the file on Synapse and
doesn't create a new version. To get around this, this function only uploads
the file to Synapse if the contents are different than what is currently on
Synapse. This function also sets the \code{set_annotations} argument of \code{synStore}
to \code{FALSE} by default so it doesn't erase any annotations.
}
\examples{
\dontrun{
filename <- "data/test_file.txt"
syn_file <- safe_synStore(filename, parent_id = "syn123", used = "syn456")
}
}
